+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+### Removed
+### Changed
+  * Enable message ordering
+  * Upgrade to `@google-cloud/pubsub#2.8.0` + UPGRADE ALL PEERS
+### Fixed
+### Deprecated
+### Security
+
+
 ## [13.6.2] -- 2021-01-11
 
 ### Fixed